more simplifications

+++ usr/sys/h/text.h
@@ -6,9 +6,9 @@
  */
 struct text
 {
-	short	x_daddr;	/* disk address of segment (relative to swplo) */
-	short	x_caddr;	/* core address, if loaded */
-	short	x_size;		/* size (clicks) */
+	int	x_daddr;	/* disk address of segment (relative to swplo) */
+	int	x_caddr;	/* core address, if loaded (Armv7: 32-bit click) */
+	int	x_size;		/* size (clicks) */
 	struct inode *x_iptr;	/* inode of prototype */
 	char	x_count;	/* reference count */
 	char	x_ccount;	/* number of loaded references */
